+++ src/include/port/win32_port.h
@@ -237,7 +237,7 @@ typedef unsigned short mode_t;
  *		Note: Some CYGWIN includes might #define WIN32.
  */
 extern int	pgsymlink(const char *oldpath, const char *newpath);
-extern int	pgreadlink(const char *path, char *buf, size_t size);
+extern ssize_t pgreadlink(const char *path, char *buf, size_t size);
 
 #define symlink(oldpath, newpath)	pgsymlink(oldpath, newpath)
 #define readlink(path, buf, size)	pgreadlink(path, buf, size)
